Easy version of code
This is the complete code. To begin with, copy and paste this code into your new sketch on the Arduino IDE. Next, look at line 257. This is our loop function which will repeat continuously until the end of times. Inside this function, I have an incomplete for loop. Try entering numbers in for VALUE to create an alternating lighting effect for different durations of time. Here is the general form of the for statement:

for (initialization; condition; iteration) 
{
// 
}
The for loop operates as follows. When the loop first starts, the initialization portion of the loop is executed. Generally, this is an expression that sets the value of the loop control variable, which acts as a counter that controls the loop. It is important to understand that the initialization expression is only executed once. Next, condition is evaluated. This must be a Boolean expression. It usually tests the loop control variable against a target value. If this expression is true, then the body of the loop is executed. If it is false, the loop terminates. Next, the iteration portion of the loop is executed. This is usually an expression that increments or decrements the loop control variable. The loop then iterates, first evaluating the conditional expression, then executing the body of the loop, and then executing the iteration expression with each pass. This process repeats until the controlling expression is false.